Avoid runtime vite dependency in server build
