+++ server/index.ts
@@ -3,7 +3,8 @@ import "./config";
 
 import express, { type Request, Response, NextFunction } from "express";
 import { registerRoutes } from "./routes";
-import { setupVite, serveStatic, log } from "./vite";
+import { serveStatic } from "./vite";
+import { log } from "./logger";
 import cors from "cors";
 import { SERVER_CONFIG } from "./config";
 
@@ -77,6 +78,7 @@ app.use((req, res, next) => {
   // setting up all the other routes so the catch-all route
   // doesn't interfere with the other routes
   if (app.get("env") === "development") {
+    const { setupVite } = await import("./vite");
     await setupVite(app, server);
   } else {
     serveStatic(app);
