+++ server/vite.ts
@@ -2,28 +2,19 @@ import { fileURLToPath } from "url";
 import express, { type Express } from "express";
 import fs from "fs";
 import path from "path";
-import { createServer as createViteServer, createLogger, type ServerOptions } from "vite";
 import { type Server } from "http";
-import viteConfig from "../vite.config";
 import { nanoid } from "nanoid";
-
-const viteLogger = createLogger();
+import type { ServerOptions } from "vite";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-export function log(message: string, source = "express") {
-  const formattedTime = new Date().toLocaleTimeString("en-US", {
-    hour: "numeric",
-    minute: "2-digit",
-    second: "2-digit",
-    hour12: true,
-  });
-
-  console.log(`${formattedTime} [${source}] ${message}`);
-}
-
 export async function setupVite(app: Express, server: Server) {
+  const { createServer: createViteServer, createLogger } = await import("vite");
+  const viteConfigPath = path.resolve(__dirname, "..", "vite.config.ts");
+  const viteConfig = (await import(viteConfigPath))?.default ?? {};
+  const viteLogger = createLogger();
+
   const serverOptions: ServerOptions = {
     middlewareMode: true,
     hmr: { server },
